Split baseline benchmarks
