+++ Sources/BaselineBenchmarks/MyComparableBenchmarks.swift.gyb
@@ -13,22 +13,32 @@
 // limitations under the License.
 
 import Benchmark
-import Foundation
 import StructuralCore
 import StructuralExamples
 
-let myDecodeJSONBenchmarks = BenchmarkSuite(name: "MyDecodeJSON") { suite in
+% for N in range(1, 17):
 
+func specializedLess(_ l: Point${N}, _ r: Point${N}) -> Bool {
+    % for i in range(1, N + 1):
+    if l._${i} == r._${i} {
+    % end
+
+    return false
+
+    % for i in range(N, 0, -1):
+    } else {
+        return l._${i} < r._${i}
+    } 
+    % end
+}
+
+% end
+
+let myComparableBenchmarks = BenchmarkSuite(name: "MyComparable") { suite in
     % for N in range(1, 17):
 
-    let inputValue${N} = [Point${N}](repeating: p${N}_1, count: 32)
-    let inputData${N} = try! JSONEncoder().encode(inputValue${N})
-    let inputString${N} = String(data: inputData${N}, encoding: .utf8)!
-
     suite.benchmark("Point${N}") {
-        var decoded: [Point${N}] = []
-        decodeJSONString(from: inputString${N}, into: &decoded)
-        anySink = decoded
+        boolSink = specializedLess(p${N}_1, p${N}_2)
     }
 
     % end
