+++ Sources/BaselineBenchmarks/MyDebugString.swift
@@ -6,94 +6,94 @@ let myDebugStringBenchmarks = BenchmarkSuite(name: "MyDebugString") { suite in
 
 
     suite.benchmark("Point1") {
-        stringSink = p1_1.debugString
+        stringSink = String(reflecting: p1_1)
     }
 
 
     suite.benchmark("Point2") {
-        stringSink = p2_1.debugString
+        stringSink = String(reflecting: p2_1)
     }
 
 
     suite.benchmark("Point3") {
-        stringSink = p3_1.debugString
+        stringSink = String(reflecting: p3_1)
     }
 
 
     suite.benchmark("Point4") {
-        stringSink = p4_1.debugString
+        stringSink = String(reflecting: p4_1)
     }
 
 
     suite.benchmark("Point5") {
-        stringSink = p5_1.debugString
+        stringSink = String(reflecting: p5_1)
     }
 
 
     suite.benchmark("Point6") {
-        stringSink = p6_1.debugString
+        stringSink = String(reflecting: p6_1)
     }
 
 
     suite.benchmark("Point7") {
-        stringSink = p7_1.debugString
+        stringSink = String(reflecting: p7_1)
     }
 
 
     suite.benchmark("Point8") {
-        stringSink = p8_1.debugString
+        stringSink = String(reflecting: p8_1)
     }
 
 
     suite.benchmark("Point9") {
-        stringSink = p9_1.debugString
+        stringSink = String(reflecting: p9_1)
     }
 
 
     suite.benchmark("Point10") {
-        stringSink = p10_1.debugString
+        stringSink = String(reflecting: p10_1)
     }
 
 
     suite.benchmark("Point11") {
-        stringSink = p11_1.debugString
+        stringSink = String(reflecting: p11_1)
     }
 
 
     suite.benchmark("Point12") {
-        stringSink = p12_1.debugString
+        stringSink = String(reflecting: p12_1)
     }
 
 
     suite.benchmark("Point13") {
-        stringSink = p13_1.debugString
+        stringSink = String(reflecting: p13_1)
     }
 
 
     suite.benchmark("Point14") {
-        stringSink = p14_1.debugString
+        stringSink = String(reflecting: p14_1)
     }
 
 
     suite.benchmark("Point15") {
-        stringSink = p15_1.debugString
+        stringSink = String(reflecting: p15_1)
     }
 
 
     suite.benchmark("Point16") {
-        stringSink = p16_1.debugString
+        stringSink = String(reflecting: p16_1)
     }
 
 
     suite.benchmark("BinaryTree") {
-        stringSink = tree1.debugString
+        stringSink = String(reflecting: tree1)
     }
 
     suite.benchmark("Color") {
-        stringSink = color1.debugString
+        stringSink = String(reflecting: color1)
     }
 
     suite.benchmark("ASCII") {
-        stringSink = ascii1.debugString
+        stringSink = String(reflecting: ascii1)
     }
 }
