+++ Sources/BaselineBenchmarks/MyEncodeJSONBenchmarks.swift.gyb
@@ -17,18 +17,13 @@ import Foundation
 import StructuralCore
 import StructuralExamples
 
-let myDecodeJSONBenchmarks = BenchmarkSuite(name: "MyDecodeJSON") { suite in
+let myEncodeJSONBenchmarks = BenchmarkSuite(name: "MyEncodeJSON") { suite in
 
     % for N in range(1, 17):
-
-    let inputValue${N} = [Point${N}](repeating: p${N}_1, count: 32)
-    let inputData${N} = try! JSONEncoder().encode(inputValue${N})
-    let inputString${N} = String(data: inputData${N}, encoding: .utf8)!
-
+    
     suite.benchmark("Point${N}") {
-        var decoded: [Point${N}] = []
-        decodeJSONString(from: inputString${N}, into: &decoded)
-        anySink = decoded
+        let data = try! JSONEncoder().encode(p${N}_1)
+        stringSink = String(data: data, encoding: .utf8)!
     }
 
     % end
