+++ Sources/BaselineBenchmarks/MyEquatableBenchmarks.swift.gyb
@@ -17,19 +17,25 @@ import Foundation
 import StructuralCore
 import StructuralExamples
 
-let myDecodeJSONBenchmarks = BenchmarkSuite(name: "MyDecodeJSON") { suite in
+let myEquatableBenchmarks = BenchmarkSuite(name: "MyEquatable") { suite in
 
     % for N in range(1, 17):
 
-    let inputValue${N} = [Point${N}](repeating: p${N}_1, count: 32)
-    let inputData${N} = try! JSONEncoder().encode(inputValue${N})
-    let inputString${N} = String(data: inputData${N}, encoding: .utf8)!
-
     suite.benchmark("Point${N}") {
-        var decoded: [Point${N}] = []
-        decodeJSONString(from: inputString${N}, into: &decoded)
-        anySink = decoded
+        boolSink = p${N}_1 == p${N}_2
     }
 
     % end
+
+    suite.benchmark("BinaryTree") {
+        boolSink = tree1 == tree2
+    }
+
+    suite.benchmark("Color") {
+        boolSink = color1 == color2
+    }
+
+    suite.benchmark("ASCII") {
+        boolSink = ascii1 == ascii2
+    }
 }
