+++ Sources/BaselineBenchmarks/MyInplaceAddBenchmarks.swift.gyb
@@ -17,18 +17,24 @@ import Foundation
 import StructuralCore
 import StructuralExamples
 
-let myDecodeJSONBenchmarks = BenchmarkSuite(name: "MyDecodeJSON") { suite in
+% for N in range(1, 17):
+
+func inplaceAddSpecialized(_ l: inout Point${N}, _ r: Point${N}) {
+    % for i in range(1, N + 1):
+    l._${i} += r._${i}
+    % end
+}
+
+% end
+
+let myInplaceAddBenchmarks = BenchmarkSuite(name: "MyInplaceAdd") { suite in
 
     % for N in range(1, 17):
 
-    let inputValue${N} = [Point${N}](repeating: p${N}_1, count: 32)
-    let inputData${N} = try! JSONEncoder().encode(inputValue${N})
-    let inputString${N} = String(data: inputData${N}, encoding: .utf8)!
-
     suite.benchmark("Point${N}") {
-        var decoded: [Point${N}] = []
-        decodeJSONString(from: inputString${N}, into: &decoded)
-        anySink = decoded
+        var p: Point${N} = p${N}_1
+        inplaceAddSpecialized(&p, p${N}_2)
+        p${N}_1 = p
     }
 
     % end
