+++ Sources/BaselineBenchmarks/SourcesSinks.swift.gyb
@@ -1,0 +1,39 @@
+import StructuralCore
+import StructuralExamples
+
+var tree1: BinaryTree<Double> = .branch(.leaf(1), 2, .branch(.leaf(3), 4, .leaf(5)))
+var tree2: BinaryTree<Double> = .branch(.leaf(10), 20, .branch(.leaf(30), 40, .leaf(50)))
+
+var color1 = Color.red
+var color2 = Color.blue
+
+var ascii1 = ASCII.tab
+var ascii2 = ASCII.lineFeed
+
+% for N in range(1, 17):
+var p${N}_2 = Point${N}(
+    % for i in range(1, N + 1):
+    _${i}: ${i}
+    % if i != N:
+    ,
+    % end
+    % end
+)
+var p${N}_1 = Point${N}(
+    % for i in range(1, N + 1):
+    _${i}: ${i} + 1
+    % if i != N:
+    ,
+    % end
+    % end
+)
+% end 
+
+var boolSink = false
+% for N in range(1, 17):
+var point${N}Sink = p${N}_1
+% end
+var treeSink = tree1
+var stringSink = ""
+var intSink: Int = 0
+var anySink: Any? = nil
