+++ Sources/StructuralBenchmarks/MyEncodeJSONBenchmarks.swift
@@ -19,148 +19,84 @@ import StructuralExamples
 
 let myEncodeJSONBenchmarks = BenchmarkSuite(name: "MyEncodeJSON") { suite in
 
-    suite.benchmark("Point1 (generic)") {
+    
+    suite.benchmark("Point1") {
         stringSink = toJSONString(p1_1)
     }
 
-    suite.benchmark("Point1 (reference)") {
-        let data = try! JSONEncoder().encode(p1_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point2 (generic)") {
+    
+    suite.benchmark("Point2") {
         stringSink = toJSONString(p2_1)
     }
 
-    suite.benchmark("Point2 (reference)") {
-        let data = try! JSONEncoder().encode(p2_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point3 (generic)") {
+    
+    suite.benchmark("Point3") {
         stringSink = toJSONString(p3_1)
     }
 
-    suite.benchmark("Point3 (reference)") {
-        let data = try! JSONEncoder().encode(p3_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point4 (generic)") {
+    
+    suite.benchmark("Point4") {
         stringSink = toJSONString(p4_1)
     }
 
-    suite.benchmark("Point4 (reference)") {
-        let data = try! JSONEncoder().encode(p4_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point5 (generic)") {
+    
+    suite.benchmark("Point5") {
         stringSink = toJSONString(p5_1)
     }
 
-    suite.benchmark("Point5 (reference)") {
-        let data = try! JSONEncoder().encode(p5_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point6 (generic)") {
+    
+    suite.benchmark("Point6") {
         stringSink = toJSONString(p6_1)
     }
 
-    suite.benchmark("Point6 (reference)") {
-        let data = try! JSONEncoder().encode(p6_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point7 (generic)") {
+    
+    suite.benchmark("Point7") {
         stringSink = toJSONString(p7_1)
     }
 
-    suite.benchmark("Point7 (reference)") {
-        let data = try! JSONEncoder().encode(p7_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point8 (generic)") {
+    
+    suite.benchmark("Point8") {
         stringSink = toJSONString(p8_1)
     }
 
-    suite.benchmark("Point8 (reference)") {
-        let data = try! JSONEncoder().encode(p8_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point9 (generic)") {
+    
+    suite.benchmark("Point9") {
         stringSink = toJSONString(p9_1)
     }
 
-    suite.benchmark("Point9 (reference)") {
-        let data = try! JSONEncoder().encode(p9_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point10 (generic)") {
+    
+    suite.benchmark("Point10") {
         stringSink = toJSONString(p10_1)
     }
 
-    suite.benchmark("Point10 (reference)") {
-        let data = try! JSONEncoder().encode(p10_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point11 (generic)") {
+    
+    suite.benchmark("Point11") {
         stringSink = toJSONString(p11_1)
     }
 
-    suite.benchmark("Point11 (reference)") {
-        let data = try! JSONEncoder().encode(p11_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point12 (generic)") {
+    
+    suite.benchmark("Point12") {
         stringSink = toJSONString(p12_1)
     }
 
-    suite.benchmark("Point12 (reference)") {
-        let data = try! JSONEncoder().encode(p12_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point13 (generic)") {
+    
+    suite.benchmark("Point13") {
         stringSink = toJSONString(p13_1)
     }
 
-    suite.benchmark("Point13 (reference)") {
-        let data = try! JSONEncoder().encode(p13_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point14 (generic)") {
+    
+    suite.benchmark("Point14") {
         stringSink = toJSONString(p14_1)
     }
 
-    suite.benchmark("Point14 (reference)") {
-        let data = try! JSONEncoder().encode(p14_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point15 (generic)") {
+    
+    suite.benchmark("Point15") {
         stringSink = toJSONString(p15_1)
     }
 
-    suite.benchmark("Point15 (reference)") {
-        let data = try! JSONEncoder().encode(p15_1)
-        stringSink = String(data: data, encoding: .utf8)!
-    }
-
-    suite.benchmark("Point16 (generic)") {
+    
+    suite.benchmark("Point16") {
         stringSink = toJSONString(p16_1)
-    }
-
-    suite.benchmark("Point16 (reference)") {
-        let data = try! JSONEncoder().encode(p16_1)
-        stringSink = String(data: data, encoding: .utf8)!
     }
 
 }
