+++ Sources/StructuralBenchmarks/MyEncodeJSONBenchmarks.swift.gyb
@@ -21,13 +21,8 @@ let myEncodeJSONBenchmarks = BenchmarkSuite(name: "MyEncodeJSON") { suite in
 
     % for N in range(1, 17):
     
-    suite.benchmark("Point${N} (generic)") {
+    suite.benchmark("Point${N}") {
         stringSink = toJSONString(p${N}_1)
-    }
-
-    suite.benchmark("Point${N} (reference)") {
-        let data = try! JSONEncoder().encode(p${N}_1)
-        stringSink = String(data: data, encoding: .utf8)!
     }
 
     % end
