+++ Sources/StructuralBenchmarks/MyEquatableBenchmarks.swift
@@ -19,155 +19,96 @@ import StructuralExamples
 
 let myEquatableBenchmarks = BenchmarkSuite(name: "MyEquatable") { suite in
 
-    suite.benchmark("Point1 (generic)") {
+
+    suite.benchmark("Point1") {
         boolSink = p1_1.myEqual(p1_2)
     }
 
-    suite.benchmark("Point1 (reference)") {
-        boolSink = p1_1 == p1_2
-    }
 
-    suite.benchmark("Point2 (generic)") {
+    suite.benchmark("Point2") {
         boolSink = p2_1.myEqual(p2_2)
     }
 
-    suite.benchmark("Point2 (reference)") {
-        boolSink = p2_1 == p2_2
-    }
 
-    suite.benchmark("Point3 (generic)") {
+    suite.benchmark("Point3") {
         boolSink = p3_1.myEqual(p3_2)
     }
 
-    suite.benchmark("Point3 (reference)") {
-        boolSink = p3_1 == p3_2
-    }
 
-    suite.benchmark("Point4 (generic)") {
+    suite.benchmark("Point4") {
         boolSink = p4_1.myEqual(p4_2)
     }
 
-    suite.benchmark("Point4 (reference)") {
-        boolSink = p4_1 == p4_2
-    }
 
-    suite.benchmark("Point5 (generic)") {
+    suite.benchmark("Point5") {
         boolSink = p5_1.myEqual(p5_2)
     }
 
-    suite.benchmark("Point5 (reference)") {
-        boolSink = p5_1 == p5_2
-    }
 
-    suite.benchmark("Point6 (generic)") {
+    suite.benchmark("Point6") {
         boolSink = p6_1.myEqual(p6_2)
     }
 
-    suite.benchmark("Point6 (reference)") {
-        boolSink = p6_1 == p6_2
-    }
 
-    suite.benchmark("Point7 (generic)") {
+    suite.benchmark("Point7") {
         boolSink = p7_1.myEqual(p7_2)
     }
 
-    suite.benchmark("Point7 (reference)") {
-        boolSink = p7_1 == p7_2
-    }
 
-    suite.benchmark("Point8 (generic)") {
+    suite.benchmark("Point8") {
         boolSink = p8_1.myEqual(p8_2)
     }
 
-    suite.benchmark("Point8 (reference)") {
-        boolSink = p8_1 == p8_2
-    }
 
-    suite.benchmark("Point9 (generic)") {
+    suite.benchmark("Point9") {
         boolSink = p9_1.myEqual(p9_2)
     }
 
-    suite.benchmark("Point9 (reference)") {
-        boolSink = p9_1 == p9_2
-    }
 
-    suite.benchmark("Point10 (generic)") {
+    suite.benchmark("Point10") {
         boolSink = p10_1.myEqual(p10_2)
     }
 
-    suite.benchmark("Point10 (reference)") {
-        boolSink = p10_1 == p10_2
-    }
 
-    suite.benchmark("Point11 (generic)") {
+    suite.benchmark("Point11") {
         boolSink = p11_1.myEqual(p11_2)
     }
 
-    suite.benchmark("Point11 (reference)") {
-        boolSink = p11_1 == p11_2
-    }
 
-    suite.benchmark("Point12 (generic)") {
+    suite.benchmark("Point12") {
         boolSink = p12_1.myEqual(p12_2)
     }
 
-    suite.benchmark("Point12 (reference)") {
-        boolSink = p12_1 == p12_2
-    }
 
-    suite.benchmark("Point13 (generic)") {
+    suite.benchmark("Point13") {
         boolSink = p13_1.myEqual(p13_2)
     }
 
-    suite.benchmark("Point13 (reference)") {
-        boolSink = p13_1 == p13_2
-    }
 
-    suite.benchmark("Point14 (generic)") {
+    suite.benchmark("Point14") {
         boolSink = p14_1.myEqual(p14_2)
     }
 
-    suite.benchmark("Point14 (reference)") {
-        boolSink = p14_1 == p14_2
-    }
 
-    suite.benchmark("Point15 (generic)") {
+    suite.benchmark("Point15") {
         boolSink = p15_1.myEqual(p15_2)
     }
 
-    suite.benchmark("Point15 (reference)") {
-        boolSink = p15_1 == p15_2
-    }
 
-    suite.benchmark("Point16 (generic)") {
+    suite.benchmark("Point16") {
         boolSink = p16_1.myEqual(p16_2)
     }
 
-    suite.benchmark("Point16 (reference)") {
-        boolSink = p16_1 == p16_2
-    }
 
-    suite.benchmark("BinaryTree (generic)") {
+    suite.benchmark("BinaryTree") {
         boolSink = tree1.myEqual(tree2)
     }
 
-    suite.benchmark("BinaryTree (reference)") {
-        boolSink = tree1 == tree2
-    }
-
-    suite.benchmark("Color (generic)") {
+    suite.benchmark("Color") {
         boolSink = color1.myEqual(color2)
     }
 
-    suite.benchmark("Color (reference)") {
-        boolSink = color1 == color2
-    }
-
-    suite.benchmark("ASCII (generic)") {
+    suite.benchmark("ASCII") {
         boolSink = ascii1.myEqual(ascii2)
-    }
-
-    suite.benchmark("ASCII (reference)") {
-        boolSink = ascii1 == ascii2
     }
 }
