+++ Sources/StructuralBenchmarks/MyEquatableBenchmarks.swift.gyb
@@ -21,37 +21,21 @@ let myEquatableBenchmarks = BenchmarkSuite(name: "MyEquatable") { suite in
 
     % for N in range(1, 17):
 
-    suite.benchmark("Point${N} (generic)") {
+    suite.benchmark("Point${N}") {
         boolSink = p${N}_1.myEqual(p${N}_2)
-    }
-
-    suite.benchmark("Point${N} (reference)") {
-        boolSink = p${N}_1 == p${N}_2
     }
 
     % end
 
-    suite.benchmark("BinaryTree (generic)") {
+    suite.benchmark("BinaryTree") {
         boolSink = tree1.myEqual(tree2)
     }
 
-    suite.benchmark("BinaryTree (reference)") {
-        boolSink = tree1 == tree2
-    }
-
-    suite.benchmark("Color (generic)") {
+    suite.benchmark("Color") {
         boolSink = color1.myEqual(color2)
     }
 
-    suite.benchmark("Color (reference)") {
-        boolSink = color1 == color2
-    }
-
-    suite.benchmark("ASCII (generic)") {
+    suite.benchmark("ASCII") {
         boolSink = ascii1.myEqual(ascii2)
-    }
-
-    suite.benchmark("ASCII (reference)") {
-        boolSink = ascii1 == ascii2
     }
 }
